fix(core): break input action initialization cycle
