+++ packages/core/src/dom/input-action-value.ts
@@ -1,0 +1,28 @@
+import { isUndefined } from '@videojs/utils/predicate';
+
+import { DEFAULT_SEEK_STEP, DEFAULT_VOLUME_STEP } from '../core/ui/constants';
+
+export type MediaInputActionName = 'seekStep' | 'volumeStep' | 'speedUp' | 'speedDown';
+
+/** Resolve the numeric value for a media input action from an explicit value or its activating key. */
+export function getMediaInputActionValue(
+  action: string,
+  key: string | undefined,
+  value?: number | undefined
+): number | undefined {
+  if (!isUndefined(value)) return value;
+
+  const normalizedKey = key?.toLowerCase();
+
+  if (action === 'seekStep') {
+    return normalizedKey === 'arrowleft' || normalizedKey === 'j' ? -DEFAULT_SEEK_STEP : DEFAULT_SEEK_STEP;
+  }
+
+  if (action === 'volumeStep') {
+    const step = DEFAULT_VOLUME_STEP / 100;
+
+    return normalizedKey === 'arrowdown' ? -step : step;
+  }
+
+  return undefined;
+}
